Move logsdb tests to logsdb plugin
